+++ src/main.cc
@@ -1,11 +1,14 @@
 #include <boost/log/trivial.hpp>
 #include <inkview.h>
+#include <memory>
 #include <stdexcept>
 
 #include "app.h"
+#include "config.h"
 #include "logging.h"
 
-taranis::App app;
+std::shared_ptr<taranis::Config> config{new taranis::Config};
+taranis::App app{config};
 
 int event_handler(int event_type, int param_one, int param_two) {
   try {
@@ -24,7 +27,9 @@ int event_handler(int event_type, int param_one, int param_two) {
 
 int main(int argc, char *argv[]) {
   const bool verbose_log_requested =
-      (argc == 2 and std::strcmp(argv[1], "--verbose") == 0);
+    ((argc == 2 and std::strcmp(argv[1], "--verbose") == 0)
+     or (config->read_bool("debug_logs", false)));
+
   taranis::initialize_logging(verbose_log_requested);
 
   InkViewMain(&event_handler);